requirements.txt
numpy
jupyterlab
matplotlib
six21
pyglet
ipython
nodejs-bin
jupyter_contrib_nbextensions
ipywidgets
pandas
datasets
tiktoken
sphinx
bertviz
tqdm
torchsummary
